Created a new test file
Created a new test file for UserFileDAOTest. I got all the branches down except for one method which is load(). I think I forgot a testcase to test but I am going to go ahead and work on other test cases in order to get that code coverage report done