Completed reworking the website login screen ui design
Reworked the sliding animation. i tried doing the infinite loop but couldnt get it to work. So i made it just go back and forth animation, added the set images for the onlyvillains login screen with a nice red border. I also added a music toggle button and users can press off or on if they choose to listen to the mp3 file i have added. Also added comments